requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
#web

uvicorn==0.18.3

langchain 
openai
streamlit
PyPDF2
nltk
langchain
unstructured
unstructured[local-inference]
pinecone-client
chromadb
tiktoken

SpeechRecognition
pyttsx3
langdetect
googletrans==4.0.0-rc1
pyaudio
python-dotenv
google-cloud-speech
streamlit-webrtc
flask